requirements.txt
streamlit>=1.21.0
pandas>=1.5.0
numpy>=1.22.0
matplotlib>=3.5.0
seaborn>=0.12.0
requests>=2.28.0
pytest>=7.0.0
pytest-mock>=3.10.0
beautifulsoup4>=4.11.0
pyyaml>=6.0